requirements.txt
arxiv
requests
python-dotenv
schedule 
tqdm